Remove TokenSecuredServices from Sample app and place in separate self-containing app. Created one for using the Challenge method and one for a known user login.
