+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredServicesChallenge/README.md
@@ -1,0 +1,8 @@
+Token Secured Services (Challenge method) - Desktop
+
+This sample demonstrates accessing secured services using token-based authentication. Token-based authentication is supported for ArcGIS Server, ArcGIS Online, and Portal for ArcGIS. The application is responsible for providing a dialog to prompt the user for credentials. With the ArcGIS Runtime SDK for .NET, you can use the IdentityManager to challenge for credentials when the Map tries to access a secure service.
+
+--------------------
+
+Additional Resources:
+https://developers.arcgis.com/net/desktop/guide/use-arcgis-token-authentication.htm
