+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredServicesKnownUser/App.xaml
@@ -1,0 +1,8 @@
+﻿<Application x:Class="TokenSecuredServices.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             StartupUri="MainWindow.xaml"
+			 Startup="Application_Startup">
+	<Application.Resources>
+	</Application.Resources>
+</Application>
