+++ src/Desktop/ArcGISRuntimeSamplesDesktop/Samples/Security/TokenSecuredServices/TokenSecuredServicesKnownUser/README.md
@@ -1,0 +1,10 @@
+Token Secured Services (Known User) - Desktop
+
+This sample demonstrates accessing secured services using token-based authentication. Token-based authentication is supported for ArcGIS Server, ArcGIS Online, and Portal for ArcGIS. 
+
+In this sample a known user is passed in, so there is no login dialog supplied to prompt the user for credentials when the secure resource is accessed. To see an example of using the ArcGIS Runtime SDK for .NET Identity Manager to prompt for credentials, see the Token Secured Services (Challenge method) sample. 
+
+--------------------
+
+Additional Resources:
+https://developers.arcgis.com/net/desktop/guide/use-arcgis-token-authentication.htm
